refactored move_to method, made castling check valid, working on castling move

+++ app/models/king.rb
@@ -1,5 +1,6 @@
 # King model inheriting from Piece
 class King < Piece
+
   def valid_move?(future_x, future_y)
     !obstructed?(future_x, future_y) && valid_king_range?(future_x, future_y)
   end
@@ -15,14 +16,8 @@ class King < Piece
 
   def move_to!(new_x, new_y)
     potential_move = game.pieces.find_by(x_coordinate: new_x, y_coordinate: new_y)
-    if potential_move.nil? 
-      cancel_castling
-      return update_attributes(x_coordinate: new_x, \
-                             y_coordinate: new_y)
-    end
-    return raise 'Illegal move.' if color == potential_move.color
-    cancel_castling
-    potential_piece.destroy
+    return raise 'Illegal move.' if !potential_move.nil? && color == potential_move.color
+    potential_piece.destroy if !potential_move.nil?
     update_attributes(x_coordinate: new_x, y_coordinate: new_y)
   end
 
@@ -46,7 +41,4 @@ class King < Piece
     return true if x_diff == 1 && y_diff == 1
   end
 
-  def cancel_castling
-    color == "white" ? game.update_attributes(white_castling: true) : game.black_castling = true
-  end
 end
